+++ server/src/main/java/tig/server/member/dto/RefreshTokenResponseDto.java
@@ -2,8 +2,6 @@ package tig.server.member.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import tig.server.kakao.dto.LoginMemberResponseDto;
-import tig.server.member.domain.Member;
 
 @Getter
 @AllArgsConstructor
