+++ server/src/main/java/tig/server/member/service/MemberService.java
@@ -10,8 +10,8 @@ import tig.server.enums.MemberRoleEnum;
 import tig.server.global.exception.BusinessExceptionHandler;
 import tig.server.global.code.ErrorCode;
 import tig.server.jwt.TokenProvider;
-import tig.server.kakao.dto.KakaoUserInfoResponseDto;
-import tig.server.kakao.dto.LoginMemberResponseDto;
+import tig.server.oauth2.kakao.dto.KakaoUserInfoResponseDto;
+import tig.server.oauth2.kakao.dto.LoginMemberResponseDto;
 import tig.server.member.domain.Member;
 import tig.server.member.dto.MemberResponse;
 import tig.server.member.dto.RefreshTokenResponseDto;
