+++ server/src/main/java/tig/server/oauth2/kakao/dto/LoginMemberResponseDto.java
@@ -1,4 +1,4 @@
-package tig.server.kakao.dto;
+package tig.server.oauth2.kakao.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
