+++ server/src/main/java/tig/server/search/dto/AvgPointDto.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import tig.server.kakao.dto.LoginAccessTokenResponseDto;
 
 @Getter
 @Setter
